BotTranslatorTelegram1/requirements.txt
psycopg2-binary==2.8.6
gunicorn==20.0.4
Flask==1.1.2
Jinja2==2.11.3
SQLAlchemy==1.4.2
pyTelegramBotAPI==3.7.6